% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afterpeak_info.R
\name{afterpeak_info.rSEIQHRF}
\alias{afterpeak_info.rSEIQHRF}
\title{Output the peak information and time of infected reducing to certain level after the peak}
\usage{
\method{afterpeak_info}{rSEIQHRF}(x, ...)
}
\description{
Output the peak information and time of infected reducing to certain level after the peak
}
