% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afterpeak_info.R
\name{afterpeak_info.rSEIQR}
\alias{afterpeak_info.rSEIQR}
\title{New S3 method for gathering information from stochastic SEIQR model after the peak}
\usage{
\method{afterpeak_info}{rSEIQR}(x)
}
\arguments{
\item{x}{the output from simulation}
}
\description{
\code{afterpeak_info} returns the peak information and time of infected reducing to certain level after the peak
}
\examples{
model1 <- rSEIQR(N0 = 100, S0 = 99, E0 = 1 ,I0 = 0, Q0 = 0, R0 = 0,  days = 100,  pars = c(1/12, 1, 0.5, 0.15, 0.04, 0, 0, 1))
afterpeak_info(model1)
}
