% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afterpeak_info.R
\name{afterpeak_info.rSIR}
\alias{afterpeak_info.rSIR}
\title{Output the peak information and time of infected reducing to certain level after the peak}
\usage{
afterpeak_info.rSIR(x, ...)
}
\description{
Output the peak information and time of infected reducing to certain level after the peak
}
