% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidmod.R
\docType{package}
\name{epidmod}
\alias{epidmod}
\title{epidmod: A package for simulating three stochastic epidemic models}
\description{
The epidmod package provides categories of important functions:
rSIR, rSEIR, rSEIQHFR and multiple S3 methods
}
