% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist.R
\name{hist.rSIR}
\alias{hist.rSIR}
\title{S3 histogram for stochastic SIR model}
\usage{
\method{hist}{rSIR}(x, y, type = c("time", "proportion"))
}
\arguments{
\item{x}{the output from simulation}

\item{y}{the input infected proportion or a certain time}

\item{type}{the type of the second argument}
}
\description{
\code{hist} returns histogram for each group given a certain a time or an infected rate for SIR model
}
\examples{
model1 <- rSIR (N0 = 1000, I0 = 0, S0 = 1000, days = 300, pars = c(1/10, 1, 1/5, 0.1, 0.1))
hist( model1, 81, type = "time" )
hist( model1, 0.7, type = "proportion" )
}
