% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.rSEIQR}
\alias{plot.rSEIQR}
\title{S3 plot method for stochastic SEIQR model}
\usage{
\method{plot}{rSEIQR}(x)
}
\arguments{
\item{x}{the output from simulation}
}
\description{
\code{plot} returns the time series plot of simulated SEIQR model
}
\examples{
model1 <- rSEIQR(N0 = 1000, I0 = 0, S0 = 999, R0 = 0, E0 = 1 , days = 300, pars = c(1/12, 1, 1/4, 0.15, 1/5, 0.3, 0.2, 0.7))
plot(model1)

}
