% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.rSIR}
\alias{plot.rSIR}
\title{S3 plot method for stochastic SIR model}
\usage{
\method{plot}{rSIR}(x)
}
\arguments{
\item{x}{the output from simulation}
}
\description{
\code{plot} returns the time series plot of simulated SIR model
}
\examples{
model1 <- rSIR (N0 = 1000, I0 = 0, S0 = 1000, days = 300, pars = c(1/10, 1, 1/5, 0.1, 0.1))
plot(model1)

}
