% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.rSEIQHRF}
\alias{print.rSEIQHRF}
\title{S3 print method for stochastic SEIQHRF model}
\usage{
\method{print}{rSEIQHRF}(x, ...)
}
\arguments{
\item{x}{the output from simulation}
}
\description{
\code{print} returns the basic information of simulated SEIQHRF model
}
\examples{
para <- c(4,2,1,2,1,3,1,2,1,2,0.9, 0.3, 0.4,0.1, 0.1)
model1 <- rSEIQHRF(N0 = 1000, S0 = 999, E0 = 1, I0 = 0, Q0 = 0, H0 = 0, R0 = 0, F0 = 0, days = 300, pars = para )
print(model1)

}
