% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.rSEIR}
\alias{print.rSEIR}
\title{S3 print method for stochastic SEIR model}
\usage{
\method{print}{rSEIR}(x)
}
\arguments{
\item{x}{the output from simulation}
}
\description{
\code{print} returns the basic information of simulated SEIR model
}
\examples{
model1 <- rSEIR(N0 = 1000, I0 = 0, S0 = 999, R0 = 0, E0 = 1 ,
                days = 300, pars = c(1/12, 1, 1/4, 1/5, 0.3, 0.2, 0.7))
print(model1)

}
