% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.rSIR}
\alias{print.rSIR}
\title{S3 print method for stochastic SIR model}
\usage{
\method{print}{rSIR}(x)
}
\arguments{
\item{x}{the output from simulation}
}
\description{
\code{print} returns the basic information of simulated SIR model
}
\examples{
model1 <- rSIR (N0 = 1000, I0 = 0, S0 = 1000,
                days = 300, pars = c(1/10, 1, 1/5, 0.1, 0.1))
print(model1)

}
