% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSEIQHRF.R
\name{rSEIQHRF}
\alias{rSEIQHRF}
\title{Simulates from a simple stochastic SEIQHRF epidemic model.}
\usage{
rSEIQHRF(
  N0 = 1000,
  S0 = 999,
  E0 = 1,
  I0 = 0,
  Q0 = 0,
  H0 = 0,
  R0 = 0,
  F0 = N0 - S0 - E0 - I0 - Q0 - H0 - R0,
  days = 100,
  pars = c(10, 2, 1, 2, 1, 3, 1, 3, 1, 1, 0.9, 0.3, 0.4, 0.1, 0.1)
)
}
\arguments{
\item{N0}{An integer. The population size at time 0}

\item{S0}{An integer. The initial number of susceptible people.}

\item{E0}{An integer. The initial number of exposed people.}

\item{I0}{An integer. The initial number of infected people.}

\item{Q0}{An integer. The initial number of quarantined people.}

\item{H0}{An integer. The initial number of people require hospitalization.}

\item{R0}{An integer. The initial number of recovered people.}

\item{F0}{An integer. The initial number of case fatality.}

\item{days}{An integer. The number of days for which to simulate.}

\item{pars}{A numeric vector: (a.rate, e.rate, i.rate, h1.rate, q.rate, h2.rate, r1.rate, r2.rate, r3.rate, f.rate, pE, pSR, pIim, pQim, pHim).
#' Description about pars
*\code{a.rate} represents rate of arrival of people into the population.
*\code{e.rate} represents individual effective contact rate.
*\code{i.rate} represents the incubation period is 1/i.rate days.
*\code{h1.rate}   represents infected individual requiring hospitalization rate at time t.
*\code{q.rate }   represents infected individual self-quarantining rate at time t.
*\code{h2.rate}   represents self-isolated individual requiring hospitalization rate at time t.
*\code{r1.rate}   represents recovery time for each infected individual is 1/ r1.rate.
*\code{r2.rate}   represents recovery time for each individual who accepts hospitalization is 1/ r2.rate.
*\code{r3.rate}   represents recovery time for each individual who is self-quarantined is is 1/ r3.rate.
*\code{f.rate }   represents individual who accepts hospitalization case fatality rate at time t.
*\code{pE}   represents probability that an arrival is exposed.
*\code{pSR}  represents probability that an exposed individual self-recover and turns susceptible.
*\code{pIim}  represents probability that an infected person is immune after recovery.
*\code{pQim}   represents probability that a quarantined person  is immune after recovery.
*\code{pHim}  represents probability that an individual who requires hospitalization is immune after recovery.}
}
\value{
A numeric matrix with 10 columns.  Row i contains the values of (t, S_t, E_t, I_t, Q_t, H_t, R_t, F_t, N_t) at time t.
}
\description{
\code{rSEIQHRF} returns a matrix contains information about time, number of susceptible people, exposed people, infected people,
recover people, total number of people.
}
\examples{
para <- c(4,2,1,2,1,3,1,2,1,2,0.9, 0.3, 0.4,0.1, 0.1)
model1 <- rSEIQHRF(N0 = 100, S0 = 99, E0 = 1, I0 = 0, Q0 = 0, H0 = 0, R0 = 0, F0 = 0, days = 100,pars = para )
model1
print(model1)

}
