% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSEIQR.R
\name{rSEIQR}
\alias{rSEIQR}
\title{Simulates from a simple stochastic SEIQR epidemic model.}
\usage{
rSEIQR(
  N0 = 1000,
  S0 = 999,
  E0 = 1,
  I0 = 0,
  Q0 = 0,
  R0 = N0 - S0 - E0 - I0 - Q0,
  days = 100,
  pars = c(3, 2, 2, 0.15, 1, 0.9, 0.3, 0.1)
)
}
\arguments{
\item{N0}{An integer. The population size at time 0.}

\item{S0}{An integer. The initial number of susceptible people.}

\item{E0}{An integer. The initial number of exposed people.}

\item{I0}{An integer. The initial number of infected people.}

\item{Q0}{An integer. The initial number of quarantined people.}

\item{R0}{An integer. The initial number of recovered people.}

\item{days}{An integer. The number of days for which to simulate.}

\item{pars}{A numeric vector: (a.rate, e.rate, i.rate, q.rate, r.rate, pE, pSR, pIm).
Description about pars
*\code{a.rate} represents rate of arrival of people into the population.
*\code{e.rate} represents individual effective contact rate.
*\code{i.rate} represents the incubation period is 1/i.rate days.
*\code{q.rate }   represents infected individual self-quarantining rate at time t.
*\code{r.rate} represents recovery time for each infected individual is 1/ r.rate.
*\code{pE}     represents probability that an arrival is exposed.
*\code{pSR}    represents probability that an exposed individual self-recover and turns susceptible.
*\code{pIm}    represents probability that an infected person is immune after recovery.}
}
\value{
A numeric matrix with 7 columns.  Row i contains the values of (t, S_t, E_t, I_t, Q_t, R_t, N_t) at time t.
}
\description{
\code{rSEIQR} returns a matrix contains information about time, number of susceptible people, exposed people, infected people,
quarantined people, recover people, total number of people.
}
\examples{
model1 <- rSEIQR(N0 = 100, S0 = 99, E0 = 1 ,I0 = 0, Q0 = 0, R0 = 0,  days = 100,  pars = c(1/12, 1, 0.5, 0.15, 0.04, 0, 0, 1))
model1
print(model1)

}
