% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSEIQR.R, R/rSEIR.R
\name{rSEIR}
\alias{rSEIR}
\title{Simulates from a simple stochastic SEIQR epidemic model.}
\usage{
rSEIR(
  N0 = 1000,
  I0 = 0,
  S0 = 999,
  R0 = 0,
  E0 = N0 - I0 - S0,
  days = 100,
  pars = c(3, 2, 2, 1, 0.9, 0.3, 0.1)
)

rSEIR(
  N0 = 1000,
  I0 = 0,
  S0 = 999,
  R0 = 0,
  E0 = N0 - I0 - S0,
  days = 100,
  pars = c(3, 2, 2, 1, 0.9, 0.3, 0.1)
)
}
\arguments{
\item{N0}{An integer. The population size at time 0.}

\item{I0}{An integer. The initial number of infected people.}

\item{S0}{An integer. The initial number of susceptible people.}

\item{R0}{An integer. The initial number of recovered people.}

\item{E0}{An integer. The initial number of exposed people.}

\item{days}{An integer. The number of days for which to simulate.}

\item{pars}{A numeric vector: (a.rate, e.rate, i.rate, r.rate, pE, pSR, pIm).
Description about pars
*\code{a.rate} represents rate of arrival of people into the population.
*\code{e.rate} represents individual effective contact rate.
*\code{i.rate} represents the incubation period is 1/i.rate days.
*\code{r.rate} represents recovery time for each infected individual is 1/ r.rate.
*\code{pE}     represents probability that an arrival is exposed.
*\code{pSR}    represents probability that an exposed individual self-recover and turns susceptible.
*\code{pIm}    represents probability that an infected person is immune after recovery.}

\item{Q0}{An integer. The initial number of quarantined people.}
}
\value{
A numeric matrix with 7 columns.  Row i contains the values of (t, S_t, E_t, I_t, Q_t, R_t, N_t) at time t.

A numeric matrix with 6 columns.  Row i contains the values of (t, S_t, E_t, I_t, R_t, N_t) at time t.
}
\description{
\code{rSEIQR} returns a matrix contains information about time, number of susceptible people, exposed people, infected people,
quarantined people, recover people, total number of people.

\code{rSEIR} returns a matrix contains information about time, number of susceptible people, exposed people, infected people,
recover people, total number of people.
}
\examples{
model1 <- rSEIQR(N0 = 100, S0 = 99, E0 = 1 ,I0 = 0,  Q0 = 0, R0 = 0,  days = 100,  pars = c(1/12, 1, 1/4, 0.15, 1/5, 0.3, 0.2, 0.7))
model1
print(model1)

model1 <- rSEIR(N0 = 100, I0 = 0, S0 = 99, R0 = 0, E0 = 1 , days = 100,  pars = c(1/12, 1, 1/4, 1/5, 0.3, 0.2, 0.7))
model1
print(model1)

}
