% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSIR.R
\name{rSIR}
\alias{rSIR}
\title{Simulates from a simple stochastic SIR epidemic model.}
\usage{
rSIR(N0, I0, S0 = N0 - I0, R0, days, pars = c(alpha, pI, beta, pR, gamma))
}
\arguments{
\item{N0}{an integer. The population size at time 0.}

\item{I0}{an integer. The initial number of infected people.}

\item{S0}{an integer. The initial number of susceptible people.}

\item{R0}{an integer. The initial number of recovered people.}

\item{days}{an integer. The number of days for which to simulate.}

\item{pars}{a numeric vector: (alpha, pI, beta, pR, gamma).}

\item{alpha}{rate of arrival of people into the population.}

\item{pI}{probability that an arrival is infected.}

\item{beta}{individual infection rate at time t is beta / Nt .}

\item{gamma}{recovery rate for each infected individual.}

\item{pR}{probability that an infected person is immune after recovery.}
}
\value{
A numeric matrix with 5 columns.  Row i contains the values of (t, S_t, I_t, R_t, N_t) immediately after transition i - 1.
}
\description{
\code{rSIR} returns a matrix contains information about time, number of susceptible people, infected people,
recover people, total number of people.
}
\examples{
rSIR <- function(N0 = 1000, I0 = 0, S0 = N0 - I0, days = 100,pars = c(1, 0.1, 2, 1, 0.9))
}
