% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSEIQHRF.R
\name{SEIQHRFjump}
\alias{SEIQHRFjump}
\title{Simulates one jump from a simple stochastic SEIQHRF epidemic model.}
\usage{
SEIQHRFjump(x, pars)
}
\arguments{
\item{x}{a numeric vector: (time, S, E, I, Q, H, R, F, N) at the previous jump.}

\item{pars}{a numeric vector: (a.rate, e.rate, i.rate, h1.rate, q.rate, h2.rate, r1.rate, r2.rate, r3.rate, f.rate, pE, pSR, pIim, pQim, pHim).}
}
\value{
a numeric vector: (time, S, E, I, Q, H, R, F, N)  immediately after the next jump.
}
\description{
\code{SEIQHRFjump} returns states of next jump given the states right now.
}
\examples{
A <- c(0,1000,0,0,0,0,0,0,1000)
B <- c(4,2,1,2,1,3,1,2,1,2,0.9, 0.3, 0.4,0.1,0.1)
SEIQHRFjump(A,B)
}
