% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSEIQR.R
\name{SEIQRjump}
\alias{SEIQRjump}
\title{Simulates one jump from a simple stochastic SEIR epidemic model.}
\usage{
SEIQRjump(x, pars)
}
\arguments{
\item{x}{A numeric vector: (time, S, E, I, Q, R, N) at the previous jump.}

\item{pars}{A numeric vector: (a.rate, e.rate, i.rate, q.rate, r.rate, pE, pSR, pIm).}
}
\value{
A numeric vector: (time, S, E, I, Q, R, N) immediately after the next jump.
}
\description{
\code{SEIQRjump} returns states of next jump given the states right now.
}
\examples{
A <- c(0,1000,0,0,0,0,1000)
B <- c(1/12, 1, 1/4, 0.15, 1/5, 0.3, 0.2, 0.7)
SEIQRjump(A,B)
}
