% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRjump.R
\name{SEIRjump}
\alias{SEIRjump}
\title{Simulates one jump from a simple stochastic SEIR epidemic model.}
\usage{
SEIRjump(x, pars)
}
\arguments{
\item{x}{a numeric vector: (time, S, E, I, R, N) at the previous jump.}

\item{pars}{a numeric vector: (a.rate, e.rate, i.rate, r.rate, pE, pSR, pIm).}
}
\value{
a numeric vector: (time, S, E, I, R, N) immediately after the next jump.
}
\description{
\code{SEIRjump} returns states of next jump given the states right now.
}
\examples{
A <- c(0,1000,0,0,0,1000)
B <- c(5,2,4,1,0.9,0.1)
SEIRjump(A,B)
}
