% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIRjump.R
\name{SIRjump}
\alias{SIRjump}
\title{Simulates one jump from a simple stochastic SIS epidemic model.}
\usage{
SIRjump(x, pars)
}
\arguments{
\item{x}{a numeric vector: (time, S, I, R, N) at the previous jump.}

\item{pars}{a numeric vector: (alpha, pI, beta, gamma, pR).}
}
\value{
a numeric vector: (time, S, I, R, N) immediately after the next jump.
}
\description{
\code{SIRjump} returns states of next jump given the states right now.
}
\examples{
A <- c(0,1000,0,0,1000)
B <- c(1,0.1,2,1,0.9)
SIRjump(A,B)
}
