% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSEIQHRF.R, R/rSEIR.R
\name{rSEIR}
\alias{rSEIR}
\title{Simulates from a simple stochastic SEIQHRF epidemic model.}
\usage{
rSEIR(
  N0 = 1000,
  I0 = 0,
  S0 = 999,
  R0 = 0,
  E0 = N0 - I0 - S0,
  days = 100,
  pars = c(3, 2, 2, 1, 0.9, 0.1)
)

rSEIR(
  N0 = 1000,
  I0 = 0,
  S0 = 999,
  R0 = 0,
  E0 = N0 - I0 - S0,
  days = 100,
  pars = c(3, 2, 2, 1, 0.9, 0.1)
)
}
\arguments{
\item{N0}{an integer. The population size at time 0.}

\item{I0}{an integer. The initial number of infected people.}

\item{S0}{an integer. The initial number of susceptible people.}

\item{R0}{an integer. The initial number of recovered people.}

\item{E0}{an integer. The initial number of exposed people.}

\item{days}{an integer. The number of days for which to simulate.}

\item{pars}{a numeric vector: (a.rate, e.rate, i.rate, r.rate, pE, pIm).
Description about pars
*\code{a.rate} represents rate of arrival of people into the population.
*\code{e.rate} represents individual being exposed rate at time t is e.rate / Nt .
*\code{i.rate} represents individual infection rate at time t is beta / Nt .
*\code{r.rate} represents recovery rate for each infected individual.
*\code{pE}     represents probability that an arrival is exposed.
*\code{pIm}    represents probability that an infected person is immune after recovery.}

\item{Q0}{an integer. The initial number of quarantined people.}

\item{H0}{an integer. The initial number of people require hospitalization.}

\item{F0}{an integer. The initial number of case fatality.}
}
\value{
A numeric matrix with 9 columns.  Row i contains the values of (t, S_t, E_t, I_t, Q_t, R_t, H_t, R_t, F_t, N_t) at time t.

A numeric matrix with 6 columns.  Row i contains the values of (t, S_t, E_t, I_t, R_t, N_t) at time t.
}
\description{
\code{rSIR} returns a matrix contains information about time, number of susceptible people, exposed people, infected people,
recover people, total number of people.

\code{rSIR} returns a matrix contains information about time, number of susceptible people, exposed people, infected people,
recover people, total number of people.
}
\examples{
rSEIR <- function(N0 = 100, I0 = 0, S0 = 99, R0 = 0, E0 = N0 - I0 - S0, days = 100,pars = c(1, 0.3, 4, 2, 2, 0.7))
rSEIR <- function(N0 = 100, I0 = 0, S0 = 99, R0 = 0, E0 = N0 - I0 - S0, days = 100,pars = c(5, 4, 2, 2, 0.7, 0.1))
}
